fix: Temporary move TestMaxPodLogsRender to run with no race
